refactor(training-facility): point the domain layer away from the components
Nine modules under lib/training-facility imported startOfDay, isInRange and
the DateRange type from DateFilter.tsx — a 'use client' React component — for
no reason other than that being where the helpers were first written. It
typechecked and it worked, and it also meant the domain layer's import graph
contained a React component.

The helpers move to lib/training-facility/date-range.ts unchanged. DateFilter
re-exports the whole surface, so its sixteen consumers compile untouched; the
nine domain modules now import the module directly. An eslint rule scoped to
lib/** forbids @/components/** imports, because the file move fixes today's
inversion and the rule is what stops the next convenient definition from
recreating it. Verified by probe: a lib file importing DateFilter fails lint.

Groundwork for the shared training-domain package (#425) — a domain package
cannot depend on a UI package — but the arrow pointed the wrong way regardless.

Refs #425.
