+++ lib/training-facility/all-cardio.ts
@@ -12,7 +12,7 @@
  */
 
 import type { CardioActivity, CardioSession } from '@/types/cardio'
-import type { DateRange } from '@/components/training-facility/shared/DateFilter'
+import type { DateRange } from '@/lib/training-facility/date-range'
 import { parseSessionDate, type SessionAvgHrPoint } from './cardio-shared'
 
 /**
