+++ lib/training-facility/period-comparison.ts
@@ -15,7 +15,7 @@ import {
   endOfDay,
   startOfDay,
   type DateRange,
-} from '@/components/training-facility/shared/DateFilter'
+} from '@/lib/training-facility/date-range'
 
 /**
  * Minimum sessions in the previous period required before a delta is
